requirements.txt
flask==3.0.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
reportlab==4.0.8
requests==2.31.0
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2
